memory_system/requirements.txt
# Core dependencies
gradio>=4.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Logging and monitoring
loguru>=0.7.0

# Memory & Learning System
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Future dependencies (Phase 2+)
# langchain>=0.1.0
# langgraph>=0.0.20
# redis>=5.0.0
# psycopg2-binary>=2.9.0
# sqlalchemy>=2.0.0
# openai>=1.0.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
